Use mapper to remove connection handling from ServerEngine

+++ lib/user/include/User.h
@@ -4,7 +4,7 @@
 #include <string>
 #include <utility>
 
-using UserId = int;
+using UserId = unsigned long;
 
 class User
 {
